requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-cloud-speech
google-genai
google-cloud-firestore